chore(hype-ui): Added metamask injector for wagmi
