+++ services/client/src/App.tsx
@@ -6,7 +6,7 @@ import { Home, Redeem, AddHypePool, Profile } from './pages';
 import styled from 'styled-components';
 import { ModalsCenter } from './containers/modals';
 import { HypeThemeType } from './theme';
-import useMetamask from './hooks/useMetamask';
+import useWallet from './hooks/useWallet';
 
 const StyledAppContainer = styled.div<{ theme: HypeThemeType }>`
   flex: 1 0 auto;
@@ -27,7 +27,7 @@ const AppWrapper = styled.div`
 
 const Root = () => {
   const isMobile = useMediaQuery({ query: `(max-width: 950px)` });
-  const { account, connect, isConnected } = useMetamask();
+  const { account, connect, isConnected } = useWallet();
 
   return (
     <AppWrapper>
