+++ services/client/src/components/connect-wallet-btn/ConnectWallet.effects.tsx
@@ -1,8 +1,8 @@
 import { ModalsActionsEnum, useModalsDispatch } from '../../context';
-import useMetamask from '../../hooks/useMetamask';
+import useWallet from '../../hooks/useWallet';
 
 export const useConnectWalletEffects = () => {
-  const { isConnected, connect, isUnMetamaskAvailable } = useMetamask();
+  const { isConnected, connect, isUnMetamaskAvailable } = useWallet();
   const dispatchModals = useModalsDispatch();
 
   const showMetamaskInfo = () => {
