+++ services/client/src/containers/profile/Rewards.tsx
@@ -1,0 +1,65 @@
+import React from 'react';
+import { useMediaQuery } from 'react-responsive';
+import Box from '../../components/styles/Box';
+import Text from '../../components/styles/Text';
+import Heading from '../../components/styles/Heading';
+import Button from 'src/components/button/Button';
+import useWallet from 'src/hooks/useWallet';
+import { formatNumber } from '../../utils';
+
+interface RewardProps {
+  rewardAmount: number;
+  onRedeem: () => void;
+}
+
+export const RewardsContainer = (props: RewardProps) => {
+  const { isConnected } = useWallet();
+  const isMobile = useMediaQuery({ query: `(max-width: 950px)` });
+  
+  return (
+    <Box
+      p={isMobile ? '1.5rem' : '4.5rem'}
+      borderRadius="1rem"
+      display="flex"
+      flexDirection="column"
+      justifyContent="space-evenly"
+      backgroundColor="greys.1"
+      marginLeft={isMobile ? 'none' : '1rem'}
+      marginRight={isMobile ? '1rem' : 'none'}
+      marginBottom="1rem"
+      minWidth='250px'
+    >
+      <Heading
+        fontSize="1.25rem"
+        fontWeight="700"
+        lineHeight="1.625rem"
+        color="black"
+        letterSpacing="-0.02em"
+      >
+        Redeem rewards
+      </Heading>
+      <Text pt="2rem" color="greys.2" fontSize="1rem" fontWeight="700">
+        Total unredeemed
+      </Text>
+      <Heading
+        fontSize="2.25rem"
+        fontWeight="700"
+        lineHeight="2.75rem"
+        color="black"
+        letterSpacing="-0.02em"
+      >
+        {isConnected ? `${formatNumber(props.rewardAmount)} TARA` : `N/A`}
+      </Heading>
+      <Button
+        variant={
+          isConnected && props.rewardAmount && props.rewardAmount > 0 ? 'primary' : 'secondary'
+        }
+        size="small"
+        onClick={props.onRedeem}
+        disabled={!isConnected}
+      >
+        Redeem all
+      </Button>
+    </Box>
+  );
+};
