+++ services/client/src/containers/profile/index.ts
@@ -1,0 +1,3 @@
+export * from './Profile';
+export * from './Rewards';
+export * from './Card';
