+++ services/client/src/hooks/useWallet.tsx
@@ -1,10 +1,18 @@
 import { useEffect, useState } from 'react';
 import { useAccount, useConnect, useDisconnect } from 'wagmi';
+import { MetaMaskConnector } from 'wagmi/connectors/metaMask';
 
-const useMetamask = () => {
+const useWallet = () => {
   const [isConnected, setIsConnected] = useState<boolean>(false);
   const [account, setAccount] = useState<string>(null);
-  const { connect } = useConnect();
+  const connector = new MetaMaskConnector({
+    options: {
+      shimChainChangedDisconnect: false,
+    },
+  });
+  const { connect } = useConnect({
+    connector,
+  });
   const { data: currentAccount } = useAccount();
   const isUnMetamaskAvailable = !((window as any).web3 || (window as any).ethereum);
   const { disconnect } = useDisconnect();
@@ -25,4 +33,4 @@ const useMetamask = () => {
   };
 };
 
-export default useMetamask;
+export default useWallet;
