+++ services/client/src/pages/Redeem/Redeem.effects.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import useMetamask from '../../hooks/useMetamask';
+import useWallet from '../../hooks/useWallet';
 import { TransactionStatus } from '../../utils';
 
 export type TransactionItem = {
@@ -12,7 +12,7 @@ export type TransactionItem = {
 export type Reward = Omit<TransactionItem, 'status'>;
 
 export const useRedeemEffects = () => {
-  const { isConnected } = useMetamask();
+  const { isConnected } = useWallet();
   const [showHistory, setShowHistory] = useState<boolean>(false);
 
   const toggleHistory = () => {
