+++ services/client/src/theme/types.ts
@@ -26,7 +26,6 @@ export interface HypeButtons {
   secondary: ButtonColor;
   success: ButtonColor;
   danger: ButtonColor;
-  neutral: ButtonColor;
 }
 
 export interface HypeCards {
